rm models_dropout torch files
